refactor(fs): extract syncPart1 dir/time detectors
