+++ core/facts/detectors/fs/syncPart1.ts
@@ -1,96 +1,3 @@
-import { hasNode, isObject } from '../utils/astHelpers';
-
 export * from './syncPart1Metadata';
 export * from './syncPart1FileOps';
-
-export const hasFsRmSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'rmSync'
-    );
-  });
-};
-
-export const hasFsMkdirSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'mkdirSync'
-    );
-  });
-};
-
-export const hasFsReaddirSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'readdirSync'
-    );
-  });
-};
-
-export const hasFsUtimesSyncCall = (node: unknown): boolean => {
-  return hasNode(node, (value) => {
-    if (value.type !== 'CallExpression') {
-      return false;
-    }
-
-    const callee = value.callee;
-    if (!isObject(callee) || callee.type !== 'MemberExpression' || callee.computed === true) {
-      return false;
-    }
-    const objectNode = callee.object;
-    const propertyNode = callee.property;
-    return (
-      isObject(objectNode) &&
-      objectNode.type === 'Identifier' &&
-      objectNode.name === 'fs' &&
-      isObject(propertyNode) &&
-      propertyNode.type === 'Identifier' &&
-      propertyNode.name === 'utimesSync'
-    );
-  });
-};
+export * from './syncPart1DirTimes';
